+++ src/pages/cadastro-marca/CadastroMarca.jsx
@@ -1,28 +1,13 @@
-import React, {useEffect} from 'react';
-import { useHistory, useParams } from 'react-router';
+import React from 'react';
 
-import {BotaoSalvar, BotaoCancelar} from '../../@material/Button'
 import Marca from "./Marca";
 import useForm from "../../shared/hooks/useForm";
-
 import CampoDeTexto from '../../shared/components/campo-de-texto/CampoDeTexto'
 import MarcaService from "../../services/MarcaService";
+import BotaoDetalhes from "../../shared/components/botao-detalhes/BotaoDetalhes";
 
 function CadastroMarca() {
     const {atualizaValor, valores, setValores} = useForm(Marca.initialValues());
-    const history = useHistory();
-    const {id} = useParams();
-
-    function cancelar() {
-        history.goBack();
-    }
-
-    useEffect(() => {
-        if (id) {
-            MarcaService.consultar(id)
-                .then(dados => setValores(dados));
-        }
-    }, [id, setValores]);
 
     return (
         <form data-testid="form" onSubmit={(event) => {
@@ -51,23 +36,11 @@ function CadastroMarca() {
                 required={true}
                 onChange={atualizaValor}
             />
-            <div className='buttonContainer'>
-                <BotaoSalvar
-                    variant="contained"
-                    color="primary"
-                    type="submit"
-                    disabled={Marca.ehModeloValido(valores)}
-                    >
-                    {id ? 'Alterar' : 'Cadastrar'}
-                </BotaoSalvar>
-
-                <BotaoCancelar
-                    variant="contained"
-                    color="secondary"
-                    onClick={cancelar}>
-                    Cancelar
-                </BotaoCancelar>
-            </div>
+            <BotaoDetalhes
+                consultarServico={MarcaService.consultar}
+                dadosConsultados={(dados) => setValores(dados)}
+                salvarDesabilidato={Marca.ehModeloValido(valores)}
+            />
         </form>
     );
 }
